Progress on lib extraction
